refactor(select): improve structure/namings
