+++ ui-parts/select/src/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_MENU_WIDTH = 100
