+++ ui-parts/select/src/item/index.ts
@@ -1,2 +1,2 @@
-export * from './item.component.js'
-export type * from './item.interfaces.js'
+export * from './component.js'
+export type * from './interfaces.js'
