+++ ui-parts/select/src/item/interfaces.ts
