+++ ui-parts/select/src/menu/menu.css.ts
@@ -7,8 +7,5 @@ export const baseMenuStyles = style({
   flexDirection: 'column',
   outline: 'none',
   padding: 0,
-  width: 100,
   height: 'auto',
-  border: '1px solid black',
-  borderRadius: 4,
 })
