+++ ui-parts/select/src/styles/shape.css.ts
@@ -1,33 +1,22 @@
-import type { SelectShape }     from '../select.interfaces.js'
-import type { SelectShapeName } from '../select.interfaces.js'
+import type { SelectShape }     from '../interfaces.js'
+import type { SelectShapeName } from '../interfaces.js'
 
 import { style }                from '@vanilla-extract/css'
 
 import { vars }                 from '@atls-ui-parts/theme'
 
-const itemCustomShapeStyles = style({
-  padding: `${vars.space.g8} ${vars.space.g16}`,
-})
+import { DEFAULT_MENU_WIDTH }   from '../constants.js'
 
-const menuCustomShapeStyles = style({
-  borderRadius: vars.radii.f8,
-})
-
-const triggerCustomShapeStyles = style({
-  margin: `${vars.space.zero} ${vars.space.g10}`,
-  borderRadius: vars.radii.f8,
+const menuDefaultShapeStyles = style({
+  width: DEFAULT_MENU_WIDTH,
+  borderRadius: vars.radii.f4,
 })
 
 export const shapeStyles = {
   default: {
     item: undefined,
-    menu: undefined,
+    menu: menuDefaultShapeStyles,
     trigger: undefined,
-  },
-  custom: {
-    item: itemCustomShapeStyles,
-    menu: menuCustomShapeStyles,
-    trigger: triggerCustomShapeStyles,
   },
 }
